Clean up for GridStatusProvider
