+++ src/features/page/assets/layout.css
@@ -35,6 +35,6 @@
     --title-spacing: 12px;
     --title-font-size: 1.8rem;
     --status-font-size: 0.68rem;
-    --arrow-size: 0.69em;
+    --arrow-size: 0.72em;
   }
 }
