+++ src/features/title/assets/month-spinbutton.css
@@ -2,7 +2,7 @@
   border-radius: 3px;
   display: flex;
   flex-direction: column;
-  margin: -0.09em 0 0 0;
+  margin: -0.1em 0 0 0;
   outline-color: transparent;
   outline-offset: 1px;
   outline-style: solid;
